changement de cast en int vers un cast en short pour prendre le bit signee

+++ src/op/ldi.c
@@ -6,7 +6,7 @@
 /*   By: vifonne <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/16 21:28:28 by vifonne           #+#    #+#             */
-/*   Updated: 2019/02/17 02:21:34 by vifonne          ###   ########.fr       */
+/*   Updated: 2019/02/17 04:46:01 by vifonne          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int	ldi(t_env *env, t_process *cur_process, unsigned char *bytes)
 			&& parse_imultitype(args[0], env, cur_process, decode.tab[0])
 			&& parse_imultitype(args[1], env, cur_process, decode.tab[1]))
 	{
-		tmp = *((int *)args[0]) + *((int *)args[1]);
+		tmp = *((short *)args[0]) + *((short *)args[1]);
 		target = get_in_arena(env, REG_SIZE, cur_process->pc + tmp % IDX_MOD);
 		if (target == NULL)
 			return (-1);
